solana*,sugar: inherit cargo-update-recipe-crates
:Release Notes:
The bbclass was merged upstream:
https://git.openembedded.org/openembedded-core/commit/?id=7636a2b8080521ed2ad54b0edce47a8742a12d58

:Detailed Notes:
Be aware that this still requires newer python3 3.11 from:
https://git.openembedded.org/openembedded-core/commit/?id=f10cdc155e47af5627ee999c57e1d083f9382a91
which wasn't backported (as it requires more changes in other recipes
and we can regenerate the list of crates with Mickledore based build).
This is just to allow the bbclass to be inherited in solana and sugar
already.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRO-12360] Prepare CCC for the solana NFT minting based on build-webos yocto 4.0
[WRO-15128] CCC: Various build fixes

Change-Id: Ie48df068a230af18b99bc1b58976c5db0a9170cf