qtbase=webos104 qtdeclarative=webos81 qtwayland=webos37
:Release Notes:
Back to hardcoded maxver for webOS specific patches.

:Detailed Notes:
The intent of having maxver for webOS specific patches is to apply them
only in the tested Qt version. Setting it to QT_MODULE_BRANCH lets those
patches applied whatever Qt version is, especially in Qt's CI system
where a newer meta-qt6 is used and it may cause some build error as they
are not integrated in the upstream.

:Testing Performed:
See the CCC jira ticket

:QA Notes:
N/A

:Issues Addressed:
[WRP-7223] CCC: qtbase=webos104 qtdeclarative=webos81 qtwayland=webos37
[WRP-7211] Fix compile issue on Qt 6.5

Change-Id: I74ece4e8334c2c9ce7f38fb5a15016624d1f7898