rpm=v4.18.0 (backport from mickledore)
:Release Notes:
We use opkg packagemanager and rpm is used only during build
for some tools like rpmdeps.

:Detailed Notes:
We don't really need the new version, but we need a fix for rpmdeps
shell environment in nativesdk from these 2 commits:
https://git.openembedded.org/openembedded-core/commit/?id=f40a2658f5be6739c5dddab7f9f11e1f85a17102
https://git.openembedded.org/openembedded-core/commit/?id=34f1121ba9684b1ab978438db2b16c1e42ea5973
and newer version was backported as-is just for simplicity of
meta-webos-backports.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-6543] CCC: Various build fixes

Change-Id: I53631a574bc06a074b3dbfb729351b5e9da6a0ef