solana-program-library: update SRCREV to get one fix from upstream
:Release Notes:
token-cli-v2.1.1 tag was missing one important fix.

:Detailed Notes:
With newer oe-core which uses --frozen instead of --offline since:
https://git.openembedded.org/openembedded-core/commit/?id=9ff9e6523bd7eb6cdc854adcbd031085c536e0e6
this was failing with:
http://gecko.lge.com:8000/Errors/Details/663185

error: the lock file TOPDIR/BUILD/work/qemux86_64-webos-linux/solana-program-library/2.1.1/git/Cargo.lock needs to be updated but --frozen was passed to prevent this
If you want to try to generate the lock file without accessing the network, remove the --frozen flag and use --offline instead.

This was already fixed in upstream in the very next commit after the tag:
commit 8177a4d9e29c8df408ac9499edea5290bddf2311
Author: hanako mumei <81144685+2501babe@users.noreply.github.com>
Date:   Wed Oct 12 13:21:31 2022 -0700

    update Cargo.lock

commit 88b147506d5b9515f3a4762421a0b8c309188dc9 (HEAD -> master, tag: token-cli-v2.1.1)
Author: hanako mumei <81144685+2501babe@users.noreply.github.com>
Date:   Tue Oct 11 19:08:42 2022 -0700

    bump token-cli to 2.1.1 and token-client to 0.2.1

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-24156] CCC: Various build fixes

Cherry-picked-from-commit: 88b71b74175df7419347958e74ed46cf8b6b8d6d
Cherry-picked-from-branch: