ose: Advanced build profile machanism
:Release Notes:

:Detailed Notes:
In the previous method, when a new profile was created, the new profile
also have to be created in the all of distro layers.
But with this patch, we can design a new profile simply inherit the
existing profile of other layers.

For example, if we introduce smartscreen profile to productize, we don't
need to create smartscreen profile for all of distro layers but only
define it to inherit webos or webos-pro's main(anything else) profile in

1) conf/distro/include/profile/webos-smartscreen.inc
    - Define profile

2) conf/distro/include/features/webos-features-smartscreen.inc
    - Extend distro features and image features.
    - Detailed configuration

:Testing Performed:
No effect on image.

:QA Notes:
N/A

:Issues Addressed:
[WRP-25790] CCC: Advanced build profile machanism
[WRP-25770] Design build profile machanism to use easily

Cherry-picked-from-commit: dd18fae8e2211a25fb30e7b20dbb602010c7838c
Cherry-picked-from-branch: