webos: webos_systemd.bbclass: Add user/group macros to support DAC
:Release Notes:
The systemd service should be run with user/group for DAC.

:Detailed Notes:
WEBOS_SYSTEMD_USER/GROUP in systemd unit files will be replaced while
install_units like other macros to avoid hard-coded User/Group.

Root is default when it's not defined or webos-dac is not in distro
features.

:Testing Performed:
Build test done.

:QA Notes:
N/A

:Issues Addressed:
[WRP-27318] CCC: Apply new DAC method
[WRP-26204] Expend package.bbclass to fix up file permissions for DAC

Cherry-picked-from-commit: b53961785426f65144aa42c06a19e126d75ab2a7
Cherry-picked-from-branch: