webos: webos_image.bbclass: Support setcap properly for webOS DAC
:Release Notes:
Add libcap-bin package when webos-dac is in distro features.

:Detailed Notes:
For webos-dac, We sometimes need to set capabilities in the binary
through the setcap command. However, because the capabilities are
written in xattr rather than the inode of the file, the attributes are
lost when the file is installed or copied.
In order to apply the capabilities in the actual installed location to
avoid above situation, we plan to use the package post install script,
which can be useful in package-level updates in the future.
When updating a package, the libcap-bin package have to be installed to
perform setcap properly on the target machine also.

:Testing Performed:
See the Jira ticket.

:QA Notes:
No effect on image.

:Issues Addressed:
[WRQ-3752] CCC: webos_image.bbclass: Support setcap
           properly for webOS DAC
[WRQ-3522] Fix to perform setcap properly for webOS DAC

Cherry-picked-from-commit: 02f9ed040a449d0f1f80c30342ad402197d76b85
Cherry-picked-from-branch: