webos: webos.inc: Add configurator to be skipped on CVE CHECK
:Release Notes:
Register configurator to be skipped on CVE CHECK

:Detailed Notes:
Configurator is an LG component.
CVE PRODUCT named oracle:configurator exists in the NVD DB,
configurator is CVE scanned.
Since the configurator recipe does not exist in oe-core, it is unlikely that the oracle:configurator is not likely to be used in the future.
Therefore, add configurator to the CVE_CHECK_SKIP_RECIPE variable to
prevent it from being CVE scanned.

:Testing Performed:
Build Tested Done.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-5315] CCC: webos: webos.inc: Add configurator to be skipped on
           CVE CHECK
[WRQ-4920] Add a recipe that doesn't check for CVEs

Cherry-picked-from-commit: 2f99748d60a4712aa287e37d87d2cf249a501955
Cherry-picked-from-branch: