webos: com.webos.service.videooutput=0 videooutput-adaptation-layer-api=0 videooutput-adaptation-layer-mock=0
:Release Notes:
Drop com.webos.service.videooutput and adaptation-layers recipes

:Detailed Notes:
videoutput is NOT being used in webOS OSE and qemul platform.
So drop the these recipes in meta-webos and they will be moved to
the meta-webos-pro layer to support other platform.

:Testing Performed:
Verified QWTC-8753 on O22
Verified QWTC-7734 on RPI4

:QA Notes:
N/A

:Issues Addressed:
[WRQ-12956] CCC: com.webos.service.videooutput=0
            videooutput-adaptation-layer-api=0
            videooutput-adaptation-layer-mock=0
[WRQ-12161] Update meta recipe and patches for the videooutput
            service if needed

Cherry-picked-from-commit: 3be6eac00394efcd3850bf306314fab05b0d0271
Cherry-picked-from-branch: