webos: webos-rpi.inc: Add hmac kernel module
:Release Notes:
Bring kernel-module-hmac explicitly.

:Detailed Notes:
Since build 2458 which includes yocto minor update for CVE and
improvement webOS quality, there are changes in meta-raspberrypi for
linux kernel as below but a regression occurred where the samba
server could not be mounted.

    c40c6e2 linux-raspberrypi: Update 5.15 recipe to 5.15.92

The root cause is that the hmac driver changed from the built-in type to
the module type. So we need to explicitly add this module on the target.

:Testing Performed:
See the JIRA ticket.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-14159] CCC: webos-rpi.inc: Add hmac kernel module
[WRQ-9441] [OSE -RPI4]: Unable to mount the samba server

Cherry-picked-from-commit: d4c854c8e09beb5c04aa8285779ee71e8f9f3847
Cherry-picked-from-branch: