webos: polkit=webos1 duktape=webos1 packagegroup-core-boot=webos4
:Release Notes:
Add polkit if it's in DISTRO_FEATURES.

:Detailed Notes:
polkit=webos1
Makes sure to have 'duktape' in PACKAGECONFIG if needed. 'duktape' is
enabled by default without PACKAGECONFIG in the current version of the
recipe but it needs to be added explicitly in a newer version as
otherwise it defaults to 'mozjs'. I think 'duktape' is sufficient for us
and better than 'mozjs' which is known to require memory much more.

duktape=webos1
Update the prebuilt binary to have missing link to libm.so.6 to address:
/usr/lib/polkit-1/polkitd: symbol lookup error: /usr/lib/libduktape.so.207: undefined symbol: __fpclassify

packagegroup-core-boot=webos4
Add polkit to webos-essential packages if it's in DISTRO_FEATURES.

:Testing Performed:
Checked if polkit is installed and running in an image with polkit added
in DISTRO_FEATURES.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-26952] CCC: polkit=webos1 duktape=webos1 packagegroup-core-boot=webos4
[WRQ-21615] [DAC] Update bootd/memchute based on polkit
[WRQ-25587] [DAC] Review bootd/memchute documents

Cherry-picked-from-commit: b5772c12ca10f246ad89c8578993fef7630850c4
Cherry-picked-from-branch: