webos: webos_image.bbclass: Add webos_checksec_scan to IMAGE_CLASSES as validitems
:Release Notes:
Introduce the checksec process in webOS

:Detailed Notes:
If webos-checksec is included in IMAGE_FEATURES, the image recipes
inherit webos_checksec_scan.
In this case, checksec is executed when each image is built.
By default, webos-checksec is not added to IMAGE_FEATURES.
It should be added as needed in external conf files such as
webos-local.conf.

:Builds Notes:
This is required by
http://gpro.lge.com/c/z-engineering/jenkins-scripts/+/420526

:Testing Performed:
Build Test Suites.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-28305] CCC: webos_image.bbclass: webos_checksec_scan to
            IMAGE_CLASSES as validitems
[WRQ-28040] Add checksec process in webOS build

Cherry-picked-from-commit: 96f7099e54c633bd0aac527e201a3d8f022e58f4
Cherry-picked-from-branch: