webos: ccache_stats.bbclass: Add compile time in ccache_stats.json
:Release Notes:
Add do_compile task elapsed time to ccache_stats.json

:Detailed Notes:
In the buildstats.bbclass, the start time of each task is recorded in
the __timedata_task variable by the set_timedata() function at the
beginning of the task.
By using the get_timedata() function of the class, the execution time of
the do_compile task can be retrieved.

Since do_ccache_stats is executed as a postfunc of do_compile, the
elapsedtime is the execution time of the do_compile task.

:Testing Performed:
Build Test Suites.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-28887] CCC: ccache_stats.bbclass: add compile time in
            ccache_stats.json
[WRQ-28864] Add compile time in ccache_stats.json

Cherry-picked-from-commit: 0bcc3a0725a5a4d679c09a2193eb138db36bab11
Cherry-picked-from-branch: