webos: Add packagegroup-core-ssh-dropbear=webos1
:Release Notes:
Drop openssh-sftp-server from
RRECOMMENDS:packagegroup-core-ssh-dropbear.

:Detailed Notes:
We used to have BAD_RECOMMENDATIONS += "openssh-sftp-server" in
webos.inc. It skips installation of openssh-sftp-server but it still
tries to build it even though it won't be used. I think it is better to
remove it from dependency chain formed by RRECOMMENDS.

:Testing Performed:
Compared webos-bom.json.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-29503] CCC: Add packagegroup-core-ssh-dropbear=webos1
[WRQ-28760] Consider static allocation of UID and GID

Cherry-picked-from-commit: 44ae968f2586e6d7ef62658b2834d80c3e843297
Cherry-picked-from-branch: