webos: Migrate Qt-related component to dynamic layer
:Release Notes:
Move Qt-related recipe of com.webos.app.test.smack.native to
qt6-layer which is dynamic layer.

:Detailed Notes:
We found some recipes that depend on qtbase and related components so
they need to be placed in meta-*/dynamic-layers/qt6-layer
We can easily monitor Qt-related modifications and also remove qt6-layer
in the future when the Qt dependency is removed.

:Testing Performed:
Local build test done.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-33319] CCC: Migrate Qt-related component to dynamic layer
[WRQ-24094] Migrate Qt-related recipes to dynamic layer

Cherry-picked-from-commit: fd2c3f3503c70918ca23b5503631129ed519e766
Cherry-picked-from-branch: