webos: com.webos.app.home=41
:Release Notes:
Home app may be suspended in the background before window.close()
is called, this causes Home app to close immediately after being
re-launched and leads to bug. It only occurs when running on the
emulator in the test system. The PR is for this specific issue.

:Detailed Notes:
submissions/39..submissions/41
9020440 Do not call window.close() when the Home app is being
        re-launched
c12d1b0 Remove the code (window.close()) that closes Home app to avoid
        causing bugs when running TCs on automation test system

:Testing Performed:
Refer CCC ticket.

:QA Notes:
N/A

:Issues Addressed:
[WRR-8131] CCC: com.webos.app.home=41
[WRR-5621] [Automation][OSE Emulator][QWTC-13533] verify logs of data
           collection on events

Cherry-picked-from-commit: 7c824910db7709e6e3fea0c2ccb2ef819e79b247
Cherry-picked-from-branch: