webos: webos-backports-5.2: add backports layer for 5.2 Walnascar
:Release Notes:
Add explicit dependency in meta-webos to make sure that
everybody executes mcf --command configure before the layer
is used will be added later, when the layer is enabled in
weboslayers.py and isn't empty.

:Detailed Notes:
This way we don't need to add the layer and update weboslayers.py at the
same time (in the same CCC).

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRR-8591] CCC: webos-backports-5.2: add backports layer for 5.2
           Walnascar

Cherry-picked-from-commit: c1ba2fe3238a22d55a59b6d1e253b3cfc46ad7c3
Cherry-picked-from-branch: