webos: Rename packagegroup-core-boot.bbappend to packagegroup-webos-core-boot.bb
:Release Notes:
Use packagegroup-webos-core-boot instead of packagegroup-core-boot.
Set RDEPENDS packagegroup-core-boot in packagegroup-webos-core-boot.

:Detailed Notes:
While proceeding with the bring-up, we encountered a case that the
image-recipe from vendor is RDEPENDING on packagegroup-core-boot.
e.g.) qti-ramdisk-image.bb from qualcomm.
Since we append the meta data for webos-core-boot using the
packagegroup-core-boot.bbappend method, unnecessary packages were
installed in the above image.

Therefore, we intend to manage the metadata for webos-core-boot in
packagegroup-webos-core-boot instead of packagegroup-core-boot.

:Testing Performed:
Build test done.

:QA Notes:
N/A

:Issues Addressed:
[WRR-9273] CCC: Rename packagegroup-core-boot.bbappend to packagegroup-
           webos-core-boot.bb
[WRR-9263] Separate packagegroup-core-boot in meta-webos to
           webOS-specific

Cherry-picked-from-commit: bf87457d808e164c7d1b5f4eccd17e630858e213
Cherry-picked-from-branch: