webos: meta-environment=webos3 (add WEBOS_SDK_BASELINE_* variables to version script)
:Release Notes:
This used to be done by upper DISTROs before, but it was broken since
2015. Now when the variables are defined in webos-sdk.inc already we
can fix it and add for all webos based DISTROs.

:Detailed Notes:
The function was set in a SDK recipe, but populate_sdk_base doesn't
inherit toolchain-scripts since 2015 Yocto 2.0 Jethro:
https://git.openembedded.org/openembedded-core/commit/?h=jethro&id=000ff33729854b1a41206a087cde1e50f49017e0

and the function wasn't called at all in SDK build, because the version
script is created by meta-environment recipe.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRR-9480] CCC: meta-environment=webos3 (add WEBOS_SDK_BASELINE_*
           variables to version script)
[WRR-2117] Create GPVB with Yocto 5.0 Scarthgap
[WRR-291] [SCM SWP] Create new toolchain(BDK) for prebuilt binaries

Cherry-picked-from-commit: ad55b0ae5686c4bda353d145792f09477e738dd4
Cherry-picked-from-branch: