webos: webos.inc: Specify the absolute path to require security_flags.inc
:Release Notes:
Set the absolute path to require security_flags.inc in webos.inc

:Detailed Notes:
In some of higher priority layers, if there exists security_flags.inc,
it will be applied.
To avoid this, in webos.inc, the security_flags.inc to be used has been
set with an absolute path.

:Testing Performed:
Build test done.

:QA Notes:
N/A

:Issues Addressed:
[WRR-10258] CCC: webos.inc: Specify the absolute path to require
            security_flags.inc
[WRR-6904] bring up webOS

Cherry-picked-from-commit: 7de1f7b42b33fe51050eacf75e5f869fb6a9718e
Cherry-picked-from-branch: