webos: webos_image.bbclass: Remove the packagegroup-core-boot in WEBOS_IMAGE_BASE_INSTALL
:Release Notes:
Remove the duplicated variable.

:Detailed Notes:
The installation of packagegroup-core-boot is executed by
packagegroup-webos-core-boot and webos_image.bbclass so that there is
a duplicated variable.

The reason for modifying webos_image.bbclass is that it makes more sense
from a webOS perspective for packagegroup-core-boot not to be included
in the WEBOS_IMAGE_BASE_INSTALL.

:Testing Performed:
Build tested.

:QA Notes:
N/A

:Issues Addressed:
[WRR-12581] CCC: webos_image.bbclass: Remove the packagegroup-core-boot
            in WEBOS_IMAGE_BASE_INSTALL
[WRR-12324] Separate packagegroup-core-boot in meta-webos to
            webOS-specific-2

Cherry-picked-from-commit: 4a344f63bc5dde2813579d299867c835fbd653b8
Cherry-picked-from-branch: