webos: com.webos.app.enactbrowser=r24
:Release Notes:
Support localization of the Enact browser when changing the system
language from the settings application.

:Detailed Notes:
In Enact Browser, there are two parts that require localization:
The main part, which includes the UIs and layout of the entire browser,
and the UI Overlay, which is used for displaying subviews, such as when
clicking the settings icon or the bookmark icon in the address bar, etc.

Currently, the resources (translated texts) are integrated for
the main part, but the UI Overlay also needs to use these resources.
To address this, the resource files will be copied to the UI
Overlay folder as well.

:Testing Performed:
1. Launch the Enact Browser app.
2. Go to https://www.google.com.
3. Open the History page.
4. Click the menu icon on the home screen of webOS OSE.
5. Change the system language via the Settings application.
6. The text displayed on the browser's UIs changes correctly.

:QA Notes:
N/A

:Issues Addressed:
[WRR-12334] CCC: com.webos.app.enactbrowser=r24
[WRR-2244] [OSE][Bug]Browser does not respond to the locale change
           via settings

Cherry-picked-from-commit: 43e9f72dcf0b7dcbdb73040ea3b5856aa721817f
Cherry-picked-from-branch: