webos: nyx-modules=r22
:Release Notes:
Remove unused RDEPENDS

:Detailed Notes:
nyx-modules:
Before uploading the coredump file to the server,
the related files are compressed.

gzip has been used for compression.
However, due to licensing issues, an old version has been used.
This time, a CVE issue occurred, and it was realistically difficult to apply it.
Therefore, we replaced it with bzip2, which does not have a licensing issue,
and at the same time, we also cleaned up RDEPENDS that are not being used.

:Testing Performed:
See CCC ticket

:QA Notes:
N/A

:Issues Addressed:
[WRR-15482] CCC: nyx-modules=r22
[WRR-13063] CVE-2022-1271 for gzip

Cherry-picked-from-commit: 01a19d24747e1b43ead69d5041bb21c379a44864
Cherry-picked-from-branch: