webos: jansson=v2.13.1
:Release Notes:
Downgrade jansson to 2.13.1 temporarily.

:Detailed Notes:
The recipe for 2.14 went in Langdale builds it using cmake which results
in all non-static symbols exported and causing a symbol clash with ulogd
for "hashtable_del". The intent is only symbols with "json_" prefix to
be exported as per
https://github.com/akheron/jansson/commit/7c707a73a2251c20afaecc028267b99d0ee60184
but the problem is that CMakeLists.txt went in later didn't respect it.
Until it gets fixed in the upstream, we choose to downgrade it to the
version that uses autotools over cmake.

:Testing Performed:
aarch64-webos-linux-nm -D libjansson.so.4  | grep hashtable_del
should get nothing.

:QA Notes:
N/A

:Issues Addressed:
[WRR-17766] CCC: jansson=v2.13.1
[WRR-15534] [Automation][Nano Mini-DQ1][QWTC-114] [Function] No failed
            systemd services
[WRR-16910] [Automation][Nano Mini-DQ1][QWTC-114] [Function] No failed
            systemd services

Cherry-picked-from-commit: 3551df5e1fcc133a6dda11498c0bba9ff3ba5f6f
Cherry-picked-from-branch: